remove duplicate state in CourseItemList and refactor filter reducer to be with search
