--- react_app/src/reducers/searchReducer.js
@@ -1,6 +1,0 @@
-export const searchString = (state = '', action) => {
-    if (action.type === 'UPDATE_SEARCH') {
-        return action.payload == null ? '' : action.payload; 
-    }
-    return state;
-}
